ancap_scraper/requirements.txt
requests
beautifulsoup4
pypdfium2
pillow
pyocr
opencv-python
pandas
scikit-learn
scikit-image
openpyxl
transformers[torch]
timm